chore(planner): trim redundant plugin-framework tests
Remove 7 test files whose coverage is duplicated by other suites:

- test_type_aware_properties.py + test_chain_augment_properties.py
  (hypothesis property tests; basic / worked-example / constrain /
  short-circuit / clamp-tracking tests already cover the algorithms)
- test_g3_real_parity.py (orchestrator G3 parity; the same scenarios
  flow through test_engine_adapter.py's test_g3_parity_via_adapter)
- test_external_plugin_wiring.py (W1+W2 startup wiring; overlaps with
  test_external_bootstrap.py + test_external_plugin_e2e.py)
- test_replay_mooncake_trace.py (real-trace replay; synthetic
  test_replay_dual_path.py already exercises dual-path replay)
- test_external_plugin_subprocess_e2e.py + subprocess runner
  (cross-process gRPC; in-process gRPC equivalent in
  test_external_plugin_e2e.py covers the same contract)

Inline the _observe_fpm_into_regressions helper into the one
remaining caller (test_builtin_plugins_e2e.py) and clean up stale
docstring references to the deleted files.
